p2p: move merkle_tree tests to tests/
Move the three tests reading block_13b8a.hex and
merkle_block.hex to tests/, along with the
get_block_13b8a helper, because they include_str data
from tests/, which we exclude when publishing.
